backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
Pillow
requests
google-generativeai
opencv-python
dlib
face_recognition
channels
channels_redis
daphne